refactor: move the error and entry classes to jvm.classes
Moves eleven classes out of the sealed `BackendObjType` hierarchy, each becoming
an object that owns its own descriptor instead of a case in the central `desc`
match. `BackendObjType.scala` drops from 2154 to 1704 lines.

New files under `phase.jvm.classes`: `GenReifiedSourceLocation`, `GenHoleError`,
`GenMatchError`, `GenCastError`, `GenUnhandledEffectError`,
`GenUncaughtExceptionHandler`, `GenMain`, and `GenNamespace`.

`Defn`, `Closure`, and `Effect` do not get their own files. They carry no fields,
methods, or bytecode — only a name — and the classes they name are emitted by
`GenFunAndClosureClasses` and `GenEffectClasses`. Their descriptors move there as
`defnDesc`, `closureDesc`, and `effectDesc`.

`Namespace(ns)` and `Main` were used as values, so their members now take the
namespace explicitly (`GenNamespace.desc(ns)`, `GenNamespace.ShimMethod(ns, defn)`).
