+++ main/test/ca/uwaterloo/flix/language/TestFlixErrors.scala
@@ -18,7 +18,7 @@ package ca.uwaterloo.flix.language
 
 import ca.uwaterloo.flix.api.{CompilerConstants, Flix}
 import ca.uwaterloo.flix.language.ast.shared.SecurityContext
-import ca.uwaterloo.flix.language.phase.jvm.BackendObjType
+import ca.uwaterloo.flix.language.phase.jvm.classes.GenHoleError
 import ca.uwaterloo.flix.runtime.{CompilationResult, JvmLoader}
 import ca.uwaterloo.flix.util.{Options, Result}
 import org.scalatest.funsuite.AnyFunSuite
@@ -33,7 +33,7 @@ class TestFlixErrors extends AnyFunSuite {
       .setOptions(Options.TestWithLibMin)
       .addVirtualPath(CompilerConstants.VirtualTestFile, input)
       .compile()
-    expectRuntimeError(result, BackendObjType.HoleError.desc.displayName())
+    expectRuntimeError(result, GenHoleError.desc.displayName())
   }
 
   test("HoleError.02") {
@@ -42,7 +42,7 @@ class TestFlixErrors extends AnyFunSuite {
       .setOptions(Options.TestWithLibMin)
       .addVirtualPath(CompilerConstants.VirtualTestFile, input)
       .compile()
-    expectRuntimeError(result, BackendObjType.HoleError.desc.displayName())
+    expectRuntimeError(result, GenHoleError.desc.displayName())
   }
 
   def expectRuntimeError(v: Result[CompilationResult, List[CompilationMessage]], name: String): Unit = {
